requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
esl==0.5.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
ply==3.11
psycopg2==2.8.4
PyJWT==1.7.1
requests==2.22.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
